Move loop to a module
