+++ switchio/__init__.py
@@ -10,7 +10,7 @@ from os import path
 from . import apps
 from .utils import get_logger, ESLError
 from .api import Client
-from .loops import get_event_loop
+from .loop import get_event_loop
 from .handlers import get_listener
 from .apps.call_gen import get_originator
 from .distribute import SlavePool, MultiEval
